Moving basefiles outside of prototype and porting Alexnet, ConvNext, Densenet and EfficientNet.
