+++ test/test_prototype_models.py
@@ -6,7 +6,7 @@ import test_models as TM
 import torch
 from common_utils import cpu_and_gpu, needs_cuda
 from torchvision.prototype import models
-from torchvision.prototype.models._api import WeightsEnum, Weights
+from torchvision.models._api import WeightsEnum, Weights
 from torchvision.prototype.models._utils import handle_legacy_interface
 
 run_if_test_with_prototype = pytest.mark.skipif(
