+++ torchvision/prototype/models/detection/faster_rcnn.py
@@ -13,8 +13,8 @@ from ....models.detection.faster_rcnn import (
     misc_nn_ops,
     overwrite_eps,
 )
-from .._api import WeightsEnum, Weights
-from .._meta import _COCO_CATEGORIES
+from torchvision.models._api import WeightsEnum, Weights
+from torchvision.models._meta import _COCO_CATEGORIES
 from .._utils import handle_legacy_interface, _ovewrite_value_param
 from ..mobilenetv3 import MobileNet_V3_Large_Weights, mobilenet_v3_large
 from ..resnet import ResNet50_Weights, resnet50
