--- torchvision/prototype/models/mobilenet.py
@@ -1,6 +1,0 @@
-from .mobilenetv2 import *  # noqa: F401, F403
-from .mobilenetv3 import *  # noqa: F401, F403
-from .mobilenetv2 import __all__ as mv2_all
-from .mobilenetv3 import __all__ as mv3_all
-
-__all__ = mv2_all + mv3_all
