+++ torchvision/prototype/transforms/__init__.py
@@ -1,5 +1,3 @@
-from torchvision.transforms import InterpolationMode, AutoAugmentPolicy  # usort: skip
-
 from . import functional  # usort: skip
 
 from ._transform import Transform  # usort: skip
@@ -10,11 +8,4 @@ from ._container import Compose, RandomApply, RandomChoice, RandomOrder
 from ._geometry import HorizontalFlip, Resize, CenterCrop, RandomResizedCrop, FiveCrop, TenCrop, BatchMultiCrop
 from ._meta import ConvertBoundingBoxFormat, ConvertImageDtype, ConvertImageColorSpace
 from ._misc import Identity, Normalize, ToDtype, Lambda
-from ._presets import (
-    ObjectDetectionEval,
-    ImageClassificationEval,
-    SemanticSegmentationEval,
-    VideoClassificationEval,
-    OpticalFlowEval,
-)
 from ._type_conversion import DecodeImage, LabelToOneHot
