+++ torchvision/transforms/__init__.py
@@ -1,2 +1,9 @@
 from .transforms import *
 from .autoaugment import *
+from ._presets import (
+    ObjectDetectionEval,
+    ImageClassificationEval,
+    SemanticSegmentationEval,
+    VideoClassificationEval,
+    OpticalFlowEval,
+)
